Web/WorspaceCopilot/raspberry-faceid-web/server/requirements.txt
Flask
pymongo
imageio
mtcnn
keras-facenet
numpy
Pillow
flask-cors
paho-mqtt